Integrate VNC server into executable
The system provided `xvfb-run` handles xvfb startup,
application execution, and xvfb teardown together in one script.
This is convenient for callers but has the drawback of not allowing
us to start serving a vnc server when the xvfb has confirmed spawning.
The first attempt at this logic launched vnc as a background script after a sleep to try to align it with xvfb bootup.
This mostly worked but was a bit inconsistent and certainly felt like a hack. Instead we opt to fork the xvfb-run
executable and integrate our vnc server, which provides callers
with the same CLI arg support that they're already using. Also deal with some SIGINT and
SIGTERM routing issues that aren't handled by xvfb out of the box.